Finished Core settings Added Core text adornment rules and properties Added many settings for adornments Refractored some code

+++ BracketPairColorizer.Settings/Settings/ITypedSettingsStore.cs
@@ -10,7 +10,7 @@
 
         long GetInt64(string name, long defaultValue);
 
-        double GetDouble(string name, long defaultValue);
+        double GetDouble(string name, double defaultValue);
 
         T GetEnum<T>(string name, T defaultValue) where T : struct;
